refactor(anyharness): move session title setters to store/titles.rs (PRO-83)
Keeps sessions.rs under the repo max-lines threshold using the store's
existing per-concern file split.
